Activate target in build_rule() rather than after parsing the whole package
