+++ src/parse/BUILD
@@ -4,7 +4,6 @@ go_library(
         "init.go",
         "internal_package.go",
         "parse_step.go",
-        "suggest.go",
     ],
     resources = ["internal.tmpl"],
     visibility = ["PUBLIC"],
@@ -28,16 +27,6 @@ go_test(
     deps = [
         ":parse",
         "//src/core",
-        "//third_party/go:testify",
-    ],
-)
-
-go_test(
-    name = "suggest_test",
-    srcs = ["suggest_test.go"],
-    resources = ["internal.tmpl"],
-    deps = [
-        ":parse",
         "//third_party/go:testify",
     ],
 )
