+++ src/parse/asp/interpreter.go
@@ -98,16 +98,21 @@ func (i *interpreter) loadBuiltinStatements(s *scope, statements []*Statement, e
 	return err
 }
 
-// interpretAll runs a series of statements in the context of the given package.
-// The first return value is for testing only.
-func (i *interpreter) interpretAll(pkg *core.Package, statements []*Statement) (s *scope, err error) {
-	s = i.scope.NewPackagedScope(pkg, 1)
-	// Config needs a little separate tweaking.
-	// Annoyingly we'd like to not have to do this at all, but it's very hard to handle
-	// mutating operations like .setdefault() otherwise.
+// interpretAll interprets all the expressions with the interntion of activating the given build label. The build label
+// will be activated as soon as it is added to the graph.
+func (i *interpreter) interpretAll(pkg *core.Package, forLabel, dependent *core.BuildLabel, forSubinclude bool, statements []*Statement) (*scope, error) {
+	s := i.scope.NewPackagedScope(pkg, 1)
+	if forLabel != nil {
+		s.parsingFor = &targetActivation{
+			label:         *forLabel,
+			dependent:     *dependent,
+			forSubinclude: forSubinclude,
+		}
+	}
+
 	s.config = i.config.Copy()
 	s.Set("CONFIG", s.config)
-	_, err = i.interpretStatements(s, statements)
+	_, err := i.interpretStatements(s, statements)
 	if err == nil {
 		s.Callback = true // From here on, if anything else uses this scope, it's in a post-build callback.
 	}
@@ -188,6 +193,13 @@ func (i *interpreter) optimiseExpressions(stmts []*Statement) {
 	})
 }
 
+// targetActivation represents a request to activate a target while parsing a package
+type targetActivation struct {
+	label         core.BuildLabel
+	dependent     core.BuildLabel
+	forSubinclude bool
+}
+
 // A scope contains all the information about a lexical scope.
 type scope struct {
 	ctx             context.Context
@@ -195,6 +207,7 @@ type scope struct {
 	state           *core.BuildState
 	pkg             *core.Package
 	subincludeLabel *core.BuildLabel
+	parsingFor      *targetActivation
 	parent          *scope
 	locals          pyDict
 	config          *pyConfig
@@ -251,6 +264,7 @@ func (s *scope) NewPackagedScope(pkg *core.Package, hint int) *scope {
 		interpreter: s.interpreter,
 		state:       s.state,
 		pkg:         pkg,
+		parsingFor:  s.parsingFor,
 		parent:      s,
 		locals:      make(pyDict, hint),
 		config:      s.config,
