+++ src/parse/asp/main/main.go
@@ -61,7 +61,7 @@ func parseFile(pkg *core.Package, p *asp.Parser, filename string) error {
 		}
 		return err
 	}
-	return p.ParseFile(pkg, filename)
+	return p.ParseFile(pkg, nil, nil, false, filename)
 }
 
 type assignment struct {
