+++ src/parse/asp/objects.go
@@ -643,6 +643,7 @@ func (f *pyFunc) Call(ctx context.Context, s *scope, c *Call) pyObject {
 	s2.config = s.config
 	s2.Set("CONFIG", s.config) // This needs to be copied across too :(
 	s2.Callback = s.Callback
+	s2.parsingFor = s.parsingFor
 	// Handle implicit 'self' parameter for bound functions.
 	args := c.Arguments
 	if f.self != nil {
