+++ src/parse/asp/parser.go
@@ -80,7 +80,7 @@ func (p *Parser) MustLoadBuiltins(filename string, contents []byte) {
 // ParseFile parses the contents of a single file in the BUILD language.
 // It returns true if the call was deferred at some point awaiting  target to build,
 // along with any error encountered.
-func (p *Parser) ParseFile(pkg *core.Package, filename string) error {
+func (p *Parser) ParseFile(pkg *core.Package, label, dependent *core.BuildLabel, forSubinclude bool, filename string) error {
 	p.limiter.Acquire()
 	defer p.limiter.Release()
 
@@ -88,8 +88,7 @@ func (p *Parser) ParseFile(pkg *core.Package, filename string) error {
 	if err != nil {
 		return err
 	}
-
-	_, err = p.interpreter.interpretAll(pkg, statements)
+	_, err = p.interpreter.interpretAll(pkg, label, dependent, forSubinclude, statements)
 	if err != nil {
 		f, _ := os.Open(filename)
 		p.annotate(err, f)
@@ -128,7 +127,7 @@ func (p *Parser) ParseReader(pkg *core.Package, r io.ReadSeeker) (bool, error) {
 	if err != nil {
 		return false, err
 	}
-	_, err = p.interpreter.interpretAll(pkg, stmts)
+	_, err = p.interpreter.interpretAll(pkg, nil, nil, false, stmts)
 	return true, err
 }
