+++ test/local_subinclude_test/test_repo/local/constant.build_defs
@@ -1,0 +1,1 @@
+CONSTANT = "constant"
