+++ test/local_subinclude_test/test_repo/local/foo/BUILD_FILE
@@ -1,0 +1,10 @@
+subinclude("//local:constant")
+
+genrule(
+    name = "output",
+    outs = ["out"],
+    cmd = "echo $CONSTANT > $OUTS",
+    env = {
+        "CONSTANT": CONSTANT,
+    },
+)
